+++ models/Bounty.ts
