+++ models/index.ts
@@ -1,3 +1,4 @@
 export * from './Contributor';
 export * from './Page';
 export * from './Space';
+export * from './Bounty';
